Simplify using an interceptor with Docker
Now it is possible to mount a folder with extra classes that are loaded
by HAPI. This makes it easy to load an interceptor JAR while using the
standard Docker image. The README explains how to use it.

Spring by default uses the WarLauncher. Now the ENTRYPOINT uses the
PropertiesLauncher, configured in such a way that it is compatible with
the WarLauncher, but adding the folder /app/extra-classes to the
loader.path.